Modal created for uploading photos. Modified user and photo models to make weights in doubles and not ints.

+++ src/main/java/edu/team9/fitconnect/model/Photo.java
@@ -28,7 +28,7 @@ public class Photo {
 
     private LocalDateTime created;
 
-    private int weight;
+    private double weight;
 
     private String userId;
